+++ eda/src/autogluon/eda/__init__.py
@@ -1,4 +1,2 @@
-from .base import *
-from .backend.base import *
 from .state import *
