+++ eda/src/autogluon/eda/analysis/__init__.py
@@ -1,4 +1,4 @@
+from .base import *
 from .composite import *
+from .dataset import *
 from .interaction import *
-from .missing import *
-from .univariate import *
