--- eda/src/autogluon/eda/backend/missing.py
@@ -1,30 +1,0 @@
-from __future__ import annotations
-
-from typing import Dict, Any
-
-import matplotlib.pyplot as plt
-import missingno as msno
-
-from ..backend.base import RenderingBackend
-from ..backend.jupyter import SimpleJupyterRenderingToolsMixin
-
-ALL = '__all__'
-
-
-class MissingStatisticsRenderer(RenderingBackend, SimpleJupyterRenderingToolsMixin):
-
-    def render(self, model: Dict[str, Any]):
-        kwargs = model['kwargs']
-        for t, ds in model['datasets'].items():
-            self.render_text(f'Missing statistics for dataset: {t}', text_type='h2')
-            if model['hint'] is not None:
-                self.render_text(model['hint'])
-            if model['chart_type'] == 'matrix':
-                msno.matrix(ds, **kwargs)
-            elif model['chart_type'] == 'bar':
-                msno.bar(ds, **kwargs)
-            elif model['chart_type'] == 'heatmap':
-                msno.heatmap(ds, **kwargs)
-            elif model['chart_type'] == 'dendrogram':
-                msno.dendrogram(ds, **kwargs)
-            plt.show()
